cloud function folder now working (#3)
Updated the folder structure to support direct glcoud cli deployments. This also means that the same code is now testable locally as the code that we deploy. This required removing explicit credential setting which instead relies on Application Default Credentials, both remotely and locally. 

Readme updated with testing and deployment snippets.